--- EntornosPGit/Fichero.txt
@@ -1,6 +1,0 @@
-1
-6
-2
-100
-102
--7
