--- EntornosPGit/nbproject/private/private.properties
@@ -1,2 +1,0 @@
-compile.on.save=true
-user.properties.file=C:\\Users\\PACO\\AppData\\Roaming\\NetBeans\\12.5\\build.properties
