--- EntornosPGit/src/entornospgit/Modificador.java
@@ -1,19 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Main.java to edit this template
- */
-package entornospgit;
-
-
-
-/**
- *
- * @author Paco
- */
-public class Modificador {
-
-    //elimina esta clase, sobra
-}
-
-
-
