+++ nbproject/private/private.properties
@@ -1,0 +1,2 @@
+compile.on.save=true
+user.properties.file=C:\\Users\\admin\\AppData\\Roaming\\NetBeans\\12.5\\build.properties
